requirements.txt
Flask==2.3.2
gunicorn==21.1.0
pytesseract==0.3.10
Pillow==9.4.0
scikit-learn==1.3.0
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
nltk==3.8.1
textblob==0.16.0
wordcloud==1.9.1
tensorflow==2.13.0
